requirements.txt
textual==8.2.4
PyYAML==6.0.3
